# public/swagger/swagger-auth-api.yaml
openapi: 3.0.3
info:
  title: API d'Authentification Laravel
  description: |
    Documentation complète de l'API d'authentification avec vérification en deux étapes.

    ## Flux d'authentification
    1. **Connexion** : L'utilisateur envoie ses identifiants (email/téléphone + mot de passe)
    2. **Vérification** : Un code à 6 chiffres est envoyé par email
    3. **Validation** : L'utilisateur entre le code pour finaliser la connexion
    4. **Token** : Un token Sanctum est généré pour les requêtes authentifiées

    ## Sécurité
    - Rate limiting sur les endpoints sensibles
    - Tokens avec expiration (10 minutes pour les codes de vérification)
    - Authentification via Laravel Sanctum
  version: 1.0.0
  contact:
    name: Support API
    email: support@example.com
  license:
    name: MIT
servers:
  - url: http://localhost:8000/api
    description: Serveur de développement
  - url: https://api.example.com
    description: Serveur de production
...
paths:
...
  /auth/password/reset/{token}:
    get:
      tags:
        - Mot de passe
      summary: Vérifier le token de réinitialisation
      description: Vérifie la validité du token de réinitialisation et retourne l'email associé
      operationId: showResetForm
      parameters:
        - name: token
          in: path
          required: true
          description: Token de réinitialisation
          schema:
            type: string
      responses:
        '200':
          description: Token valide
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  email:
                    type: string
                    example: "utilisateur@example.com"
        '404':
          description: Token invalide ou expiré
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "Lien invalide ou expiré."

    post:
      tags:
        - Mot de passe
      summary: Réinitialiser le mot de passe
      description: Réinitialise le mot de passe de l'utilisateur avec le token fourni
      operationId: resetPassword
      parameters:
        - name: token
          in: path
          required: true
          description: Token de réinitialisation
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - password
                - password_confirmation
              properties:
                password:
                  type: string
                  format: password
                  minLength: 8
                  description: Nouveau mot de passe
                  example: "NouveauMotDePasse123!"
                password_confirmation:
                  type: string
                  format: password
                  description: Confirmation du nouveau mot de passe
                  example: "NouveauMotDePasse123!"
      responses:
        '200':
          description: Mot de passe réinitialisé avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Mot de passe réinitialisé avec succès."
        '404':
          description: Token invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation échouée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '500':
          description: Erreur serveur
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                  error:
                    type: string
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token Sanctum obtenu après connexion réussie.

        Format: `Bearer {token}`

        Exemple: `Bearer 1|abcdefghijklmnopqrstuvwxyz123456789`
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Message d'erreur"
    ValidationError:
      type: object
      properties:
        message:
          type: string
          example: "Les données fournies sont invalides."
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          example:
            email:
              - "Le champ email est obligatoire."
            password:
              - "Le mot de passe doit contenir au moins 8 caractères."
...